src/crewai/requirements.txt
crewai
python-dotenv
pandas
langchain_groq
groq
tiktoken
